Adds a RandomContext for generating random strings in steps.
```
Given I am viewing an "Article" with the title "<title>"
Then I should see the header "<title>"
```

When the `RandomContext` is included, then `<title>` will be replaced
with a random string during the scenario. It will have a consistent
value throughout a given scenario, but a new value in subsequent
scenarios.

- Fixes #293